pyproject.toml
[project]
name = "project-web-sge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "argon2-cffi>=25.1.0",
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.11",
    "redis>=7.0.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "flake8>=7.3.0",
    "werkzeug>=3.1.3",
]
